@@ -1,6 +1,6 @@
 {
   "name": "@aliceo2/bookkeeping",
-  "version": "0.74.0",
+  "version": "0.75.0",
   "author": "ALICEO2",
   "scripts": {
     "coverage": "nyc npm test && npm run coverage:report",
@@ -30,7 +30,7 @@
     "multer": "1.4.5-lts.1",
     "node-fetch": "3.3.1",
     "sequelize": "6.35.0",
-    "umzug": "3.4.0"
+    "umzug": "3.5.0"
   },
   "files": [
     "lib/",
@@ -40,12 +40,12 @@
   "devDependencies": {
     "@types/d3": "7.4.0",
     "chai": "4.3.4",
-    "eslint": "8.55.0",
+    "eslint": "8.56.0",
     "js-yaml": "4.1.0",
     "mocha": "10.2.0",
     "nodemon": "3.0.1",
     "nyc": "15.1.0",
-    "puppeteer": "21.5.2",
+    "puppeteer": "21.6.1",
     "puppeteer-to-istanbul": "1.4.0",
     "sequelize-cli": "6.6.0",
     "sinon": "17.0.1",
