@@ -11,12 +11,14 @@
  * or submit itself to any jurisdiction.
  */
 
-const OverviewSuite = require('./overview.test');
-const DetailSuite = require('./detail.test');
-const RunsPerPeriodOverviewSuite = require('./runsPerPeriod.overview.test');
+import { switchCase } from '/js/src/index.js';
 
-module.exports = () => {
-    describe('Overview Page', OverviewSuite);
-    describe('Detail Page', DetailSuite);
-    describe('Runs Per Period Overview Page', RunsPerPeriodOverviewSuite);
-};
+/**
+ * Formatt alice current value
+ * @param {'NEGATIVE'|'POSITIVE'|null} polarity polarity of a current
+ * @param {number} current absolute value of a current
+ * @return {string} formatted value
+ */
+export const formatAliceCurrent = (polarity, current) => (
+    switchCase(polarity, { NEGATIVE: -1, POSITIVE: 1, null: 1 }) * current
+).toLocaleString('en-US');
