@@ -14,27 +14,38 @@
 const Sequelize = require('sequelize');
 
 module.exports = (sequelize) => {
-    const LhcPeriod = sequelize.define(
-        'LhcPeriod',
+    const DataPass = sequelize.define(
+        'DataPass',
         {
             id: {
-                allowNull: false,
-                autoIncrement: true,
-                primaryKey: true,
                 type: Sequelize.INTEGER,
+                allowNull: false,
+                primaryKey: true,
+                autoIncrement: true,
             },
             name: {
                 type: Sequelize.STRING,
                 allowNull: false,
             },
+            description: {
+                type: Sequelize.TEXT,
+            },
+            outputSize: {
+                type: Sequelize.BIGINT,
+            },
+            reconstructedEventsCount: {
+                type: Sequelize.INTEGER,
+            },
+            lastRunNumber: {
+                type: Sequelize.INTEGER,
+            },
         },
-        { tableName: 'lhc_periods' },
+        { tableName: 'data_passes' },
     );
 
-    LhcPeriod.associate = (models) => {
-        LhcPeriod.hasMany(models.Run, { foreignKey: 'lhcPeriodId', as: 'runs' });
-        LhcPeriod.hasOne(models.DataPass);
+    DataPass.associate = (models) => {
+        DataPass.belongsTo(models.LhcPeriod, { foreignKey: 'lhcPeriodId', as: 'lhcPeriod' });
     };
 
-    return LhcPeriod;
+    return DataPass;
 };
