@@ -11,12 +11,13 @@
  * or submit itself to any jurisdiction.
  */
 
-const OverviewSuite = require('./overview.test');
-const DetailSuite = require('./detail.test');
-const RunsPerPeriodOverviewSuite = require('./runsPerPeriod.overview.test');
-
-module.exports = () => {
-    describe('Overview Page', OverviewSuite);
-    describe('Detail Page', DetailSuite);
-    describe('Runs Per Period Overview Page', RunsPerPeriodOverviewSuite);
-};
+/**
+ * @typedef SequelizeDataPass
+ *
+ * @property {number} id
+ * @property {string} name
+ * @property {string} description
+ * @property {number} outputSize
+ * @property {number} reconstructedEventsCount
+ * @property {number} lastRunNumber
+ */
