@@ -29,7 +29,7 @@ export const identifierColumnAndActionsDropdown = (page, linkContent, parameters
     h(
         '.flex-row.items-center.btn-group',
         [
-            frontLink(linkContent, page, parameters, { class: 'btn btn-primary' }),
+            frontLink(linkContent, page, parameters, { class: 'btn btn-primary white' }),
             dropdown(
                 h('.btn.btn-group-item.last-item', iconCaretBottom()),
                 h(
