@@ -47,7 +47,7 @@ export const logTreeViewComponent = (treeViewModel, authenticationToken) => {
                     isCollapsed,
                     () => treeViewModel.collapseLog(log.id),
                     () => treeViewModel.expandLog(log.id),
-                    authenticationToken,
+                    { authenticationToken },
                 ),
             );
 
