@@ -47,6 +47,8 @@ import { RunDefinition } from '../../../domain/enums/RunDefinition.js';
 import { coloredCalibrationStatusComponent } from '../coloredCalibrationStatusComponent.js';
 import { infoLoggerButtonGroup } from '../../../components/common/selection/infoLoggerButtonGroup/infoLoggerButtonGroup.js';
 import { BeamModes } from '../../../domain/enums/BeamModes.js';
+import { formatRunDetectorsInline } from '../format/formatRunDetectorsInline.js';
+import { formatAliceCurrent } from '../format/formatAliceCurrent.js';
 
 /**
  * List of active columns for a generic runs table
@@ -60,16 +62,19 @@ export const runsActiveColumns = {
     runNumber: {
         name: 'Run',
         visible: true,
-        size: 'w-10 f6 w-wrapped',
+        classes: 'w-10 f6 w-wrapped',
         filter: runNumberFilter,
-        format: (_, run) => infoLoggerButtonGroup(run),
+        format: (runNumber) => infoLoggerButtonGroup({ runNumber }),
         profiles: {
             lhcFill: true,
             environment: true,
+            runsPerLhcPeriod: {
+                classes: 'w-7 f6 w-wrapped',
+            },
             home: {
                 name: 'Number',
                 format: null,
-                size: 'w-20',
+                classes: 'w-20',
             },
         },
     },
@@ -78,12 +83,7 @@ export const runsActiveColumns = {
         visible: true,
         profiles: [profiles.none, 'lhcFill'],
         size: 'w-15 f6',
-        format: (detectors, run) => {
-            if (detectors && detectors.length > 0) {
-                return [h('.badge.bg-gray-light.mh2.nDetectors-badge', run.nDetectors), `${detectors.toString()}`];
-            }
-            return '-';
-        },
+        format: (_, run) => formatRunDetectorsInline(run.detectorsQualities, run.nDetectors),
         filter: detectorsFilterComponent,
         balloon: true,
     },
@@ -91,7 +91,7 @@ export const runsActiveColumns = {
         name: 'Tags',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-5 f6',
+        classes: 'w-5 f6',
         format: (tags) => tags && tags.length > 0 ? tags.map(({ text }) => text).join(', ') : '-',
         exportFormat: (tags) => tags && tags.length > 0 ? tags.map(({ text }) => text).join(', ') : '-',
         filter: (runModel) => tagFilter(runModel.listingTagsFilterModel),
@@ -100,8 +100,15 @@ export const runsActiveColumns = {
     fillNumber: {
         name: 'Fill No.',
         visible: true,
-        profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-5 f6',
+        profiles: {
+            [profiles.none]: true,
+            lhcFill: true,
+            environment: true,
+            runsPerLhcPeriod: {
+                classes: 'w-2 f6',
+            },
+        },
+        classes: 'w-5 f6',
         format: (fill, run) => fill !== null && run.lhcBeamMode === BeamModes.STABLE_BEAMS
             ? frontLink(fill, 'lhc-fill-details', { fillNumber: fill }) : '-',
         filter: fillNumbersFilter,
@@ -110,16 +117,16 @@ export const runsActiveColumns = {
         name: 'LHC Period',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-5 f6',
+        classes: 'w-5 f6',
         format: (lhcPeriod, run) => lhcPeriod && run.lhcBeamMode === BeamModes.STABLE_BEAMS ? lhcPeriod : '-',
         filter: lhcPeriodsFilter,
     },
     timeO2Start: {
         name: 'Start',
         visible: true,
-        size: 'w-7 f6',
+        classes: 'w-7 f6',
         noEllipsis: true,
-        format: (timestamp, run) => formatRunStart(run, false),
+        format: (_, run) => formatRunStart(run, false),
         exportFormat: (timestamp) => formatTimestamp(timestamp),
         filter: o2startFilter,
         profiles: {
@@ -127,17 +134,20 @@ export const runsActiveColumns = {
             environment: true,
             home: {
                 name: 'Start',
-                size: 'w-30',
+                classes: 'w-30',
                 format: (timestamp, run) => formatRunStart(run, true),
+            },
+            runsPerLhcPeriod: {
+                classes: 'f6 w-wrapped',
             },
         },
     },
     timeO2End: {
         name: 'Stop',
         visible: true,
-        size: 'w-7 f6',
+        classes: 'w-7 f6',
         noEllipsis: true,
-        format: (timestamp, run) => formatRunEnd(run, false),
+        format: (_, run) => formatRunEnd(run, false),
         exportFormat: (timestamp) => formatTimestamp(timestamp),
         filter: o2endFilter,
         profiles: {
@@ -145,31 +155,45 @@ export const runsActiveColumns = {
             environment: true,
             home: {
                 name: 'Stop',
-                size: 'w-30',
-                format: (timestamp, run) => formatRunEnd(run, true),
-
+                classes: 'w-30',
+                format: (_, run) => formatRunEnd(run, true),
+            },
+            runsPerLhcPeriod: {
+                classes: 'f6 w-wrapped',
             },
         },
     },
     timeTrgStart: {
         name: 'TRG Start',
         visible: false,
-        size: 'w-10 f6 w-wrapped',
+        classes: 'w-10 f6 w-wrapped',
         format: (timestamp) => formatTimestamp(timestamp, false),
         exportFormat: (timestamp) => formatTimestamp(timestamp),
+        profiles: {
+            runsPerLhcPeriod: {
+                visible: true,
+                classes: 'f6 w-wrapped',
+            },
+        },
     },
     timeTrgEnd: {
         name: 'TRG Stop',
         visible: false,
-        size: 'w-10 f6 w-wrapped',
+        classes: 'w-10 f6 w-wrapped',
         format: (timestamp) => formatTimestamp(timestamp, false),
         exportFormat: (timestamp) => formatTimestamp(timestamp),
+        profiles: {
+            runsPerLhcPeriod: {
+                visible: true,
+                classes: 'f6 w-wrapped',
+            },
+        },
     },
     timeSincePreviousRun: {
         name: 'Since prev.',
         visible: true,
         profiles: ['lhcFill'],
-        size: 'f6 w-10',
+        classes: 'f6 w-10',
         format: (duration) => duration !== null
             ? formatDuration(duration)
             : h('strong.danger', 'UNKNOWN'),
@@ -178,7 +202,7 @@ export const runsActiveColumns = {
         name: 'Definition',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-10 f6 w-wrapped',
+        classes: 'w-10 f6 w-wrapped',
         format: (definition, run) => {
             const lines = [h('', definition || '-')];
             if (definition === RunDefinition.Calibration) {
@@ -191,7 +215,7 @@ export const runsActiveColumns = {
     runDuration: {
         name: 'Duration',
         visible: true,
-        size: 'w-7 f6',
+        classes: 'w-7 f6',
         profiles: [profiles.none, 'lhcFill', 'environment'],
         noEllipsis: true,
         format: (_duration, run) => displayRunDuration(run),
@@ -202,14 +226,14 @@ export const runsActiveColumns = {
         name: 'Environment ID',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-10 f6 w-wrapped',
+        classes: 'w-10 f6 w-wrapped',
         filter: environmentIdFilter,
         format: (id) => id ? frontLink(id, 'env-details', { environmentId: id }) : '-',
     },
     runType: {
         name: 'Run Type',
         visible: false,
-        size: 'cell-l f6 w-wrapped',
+        classes: 'cell-l f6 w-wrapped',
         format: formatRunType,
         filter: (runModel) => selectionDropdown(runModel.listingRunTypesFilterModel, { selectorPrefix: 'run-types' }),
     },
@@ -217,20 +241,20 @@ export const runsActiveColumns = {
         name: 'Quality',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-5 f6 w-wrapped',
+        classes: 'w-5 f6 w-wrapped',
         filter: runQualityFilter,
     },
     nDetectors: {
         name: 'DETs #',
         visible: false,
-        size: 'w-2 f6 w-wrapped',
+        classes: 'w-2 f6 w-wrapped',
         filter: nDetectorsFilter,
     },
     nEpns: {
         name: 'EPNs #',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-2 f6 w-wrapped',
+        classes: 'w-2 f6 w-wrapped',
         // eslint-disable-next-line no-extra-parens
         format: (nEpns, run) => run.epn ? (typeof nEpns === 'number' ? nEpns : 'ON') : 'OFF',
         filter: nEpnsFilter,
@@ -239,24 +263,24 @@ export const runsActiveColumns = {
         name: 'FLPs #',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-2 f6 w-wrapped',
+        classes: 'w-2 f6 w-wrapped',
         filter: nFlpsFilter,
     },
     nSubtimeframes: {
         name: '# of STFs',
         visible: false,
-        size: 'cell-s f6 w-wrapped',
+        classes: 'cell-s f6 w-wrapped',
     },
     bytesReadOut: {
         name: 'Readout Data',
         visible: false,
-        size: 'cell-m f6 w-wrapped',
+        classes: 'cell-m f6 w-wrapped',
     },
     dd_flp: {
         name: 'Data Distribution (FLP)',
         visible: false,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-2 f6 w-wrapped',
+        classes: 'w-2 f6 w-wrapped',
         format: (boolean) => boolean ? 'On' : 'Off',
         exportFormat: (boolean) => boolean ? 'On' : 'Off',
         filter: ddflpFilter,
@@ -265,7 +289,7 @@ export const runsActiveColumns = {
         name: 'DCS',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-2 f6 w-wrapped',
+        classes: 'w-2 f6 w-wrapped',
         format: (boolean) => boolean ? 'On' : 'Off',
         exportFormat: (boolean) => boolean ? 'On' : 'Off',
         filter: dcsFilter,
@@ -274,7 +298,7 @@ export const runsActiveColumns = {
         name: 'TRG',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-5 f6 w-wrapped',
+        classes: 'w-5 f6 w-wrapped',
         filter: triggerValueFilter,
         format: (trgValue) => trgValue ? trgValue : '-',
     },
@@ -282,7 +306,7 @@ export const runsActiveColumns = {
         name: 'EPN',
         visible: false,
         profiles: [profiles.none, 'lhcFill', 'environment'],
-        size: 'w-2 f6 w-wrapped',
+        classes: 'w-2 f6 w-wrapped',
         format: (boolean) => boolean ? 'On' : 'Off',
         exportFormat: (boolean) => boolean ? 'On' : 'Off',
         filter: epnFilter,
@@ -290,12 +314,12 @@ export const runsActiveColumns = {
     epnTopology: {
         name: 'EPN Topology',
         visible: false,
-        size: 'w-10 f6',
+        classes: 'w-10 f6',
         balloon: true,
     },
     odcTopologyFullName: {
         name: 'Topology Full',
-        size: 'w-15 f6',
+        classes: 'w-15 f6',
         visible: false,
         profiles: [profiles.none, 'lhcFill', 'environment'],
         filter: odcTopologyFullName,
@@ -303,7 +327,7 @@ export const runsActiveColumns = {
     },
     eorReasons: {
         name: 'EOR Reason',
-        size: 'w-15 f6',
+        classes: 'w-15 f6',
         visible: true,
         profiles: [profiles.none, 'lhcFill', 'environment'],
         format: (eorReasons) => displayRunEorReasonsOverview(eorReasons),
@@ -360,7 +384,7 @@ export const runsActiveColumns = {
         name: '',
         visible: true,
         profiles: 'home',
-        size: 'w-15',
+        classes: 'w-15',
         format: (_, run) => h(
             '.flex-row',
             { style: 'justify-content: end' },
@@ -371,5 +395,18 @@ export const runsActiveColumns = {
                 { id: `btn${run.runNumber}`, class: 'btn btn-primary btn-sm btn-redirect' },
             ),
         ),
+    },
+    aliceL3Current: {
+        name: 'L3 [A]',
+        visible: true,
+        format: (_, run) => formatAliceCurrent(run.aliceL3Polarity, run.aliceL3Current),
+        profiles: ['runsPerLhcPeriod'],
+    },
+
+    dipoleCurrent: {
+        name: 'Dipole [A]',
+        visible: true,
+        format: (_, run) => formatAliceCurrent(run.aliceDipolePolarity, run.aliceDipoleCurrent),
+        profiles: ['runsPerLhcPeriod'],
     },
 };
