@@ -30,8 +30,7 @@ import pick from '../../../utilities/pick.js';
 export class RunsOverviewModel extends Observable {
     /**
      * The constructor of the Overview model object
-     * @param {Model} model Pass the model to access the defined functions
-     * @return {Object} Constructs the Overview model
+     * @param {Model} model global model
      */
     constructor(model) {
         super();
@@ -61,9 +60,8 @@ export class RunsOverviewModel extends Observable {
         this._allRuns = RemoteData.notAsked();
 
         this.reset(false);
-
         // eslint-disable-next-line no-return-assign,require-jsdoc
-        const updateDebounceTime = () => this._debouncedFetchAllRuns = debounce(() => this.fetchRuns(), model.inputDebounceTime);
+        const updateDebounceTime = () => this._debouncedFetchAllRuns = debounce(this.fetchRuns.bind(this), model.inputDebounceTime);
         model.appConfiguration$.observe(() => updateDebounceTime());
         updateDebounceTime();
     }
