@@ -12,14 +12,14 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
-import { runsActiveColumns } from '../ActiveColumns/runsActiveColumns.js';
-import { exportRunsModal } from './exportRunsModal.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
-import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
+import { exportRunsTriggerAndModal } from './exportRunsTriggerAndModal.js';
 import { filtersPanelPopover } from '../../../components/Filters/common/filtersPanelPopover.js';
+import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
+import { runsActiveColumns } from '../ActiveColumns/runsActiveColumns.js';
+import { showPhysicsButton } from './showPhysicsButton.js';
+import { table } from '../../../components/common/table/table.js';
 import runNumberFilter from '../../../components/Filters/RunsFilter/runNumber.js';
-import { RunDefinition } from '../../../domain/enums/RunDefinition.js';
 
 const TABLEROW_HEIGHT = 59;
 // Estimate of the navbar and pagination elements height total; Needs to be updated in case of changes;
@@ -39,43 +39,13 @@ export const RunsOverviewPage = ({ runs: { overviewModel: runsOverviewModel }, m
     return h('', [
         h('.flex-row.header-container.g2.pv2', [
             filtersPanelPopover(runsOverviewModel, runsActiveColumns),
+            h('.pl2#runOverviewFilter', runNumberFilter(runsOverviewModel)),
             showPhysicsButton(runsOverviewModel),
-            h('.w-20#runOverviewFilter', runNumberFilter(runsOverviewModel)),
-            exportRunsButton({ runsOverviewModel, modalModel }),
+            exportRunsTriggerAndModal(runsOverviewModel, modalModel),
         ]),
         h('.flex-column.w-100', [
             table(runsOverviewModel.runs, runsActiveColumns),
             paginationComponent(runsOverviewModel.pagination),
         ]),
     ]);
-};
-
-/**
- * Builds a button which will redirect the user to the export run page
- * @param {Model} model global model to access functions
- * @returns {vnode} with button
- */
-const exportRunsButton = ({ runsOverviewModel, modalModel }) => h('button.btn.btn-primary.w-15.h2.mlauto#export-runs-trigger', {
-    disabled: (runsOverviewModel.runs?.payload?.length ?? 0) === 0,
-    onclick: () => modalModel.display({ content: (modalModel) => exportRunsModal(runsOverviewModel, modalModel), size: 'medium' }),
-}, 'Export Runs');
-
-/**
- * Builds a button which will toggle the physics filter
- * @param {OverviewModel} runsOverviewModel the model of the runs overview
- * @returns {vnode} with button
- */
-const showPhysicsButton = (runsOverviewModel) => {
-    const isPhysicsActive = runsOverviewModel.isDefinitionInFilter(RunDefinition.Physics);
-    const buttonColor = isPhysicsActive ? 'primary' : 'secondary';
-
-    return h(`button.btn.btn-${buttonColor}.h2#show-physics-trigger`, {
-        onclick: () => {
-            if (isPhysicsActive) {
-                runsOverviewModel.removeDefinitionFilter(RunDefinition.Physics);
-            } else {
-                runsOverviewModel.addDefinitionFilter(RunDefinition.Physics);
-            }
-        },
-    }, 'Show PHYSICS');
 };
