@@ -92,11 +92,11 @@ const errorDisplay = () => h('.danger', 'Data fetching failed');
 
 /**
  * A function to construct the exports runs screen
- * @param {Model} runsOverviewModel Pass the model to access the defined functions
+ * @param {RunsOverviewModel} runsOverviewModel Pass the model to access the defined functions
  * @param {ModalHandler} modalHandler The modal handler, used to dismiss modal after export
- * @return {vnode} Return the view of the inputs
+ * @return {Component} Return the view of the inputs
  */
-export const exportRunsModal = (runsOverviewModel, modalHandler) => {
+const exportRunsModal = (runsOverviewModel, modalHandler) => {
     const runsRemoteData = runsOverviewModel.allRuns;
 
     return h('div#export-runs-modal', [
@@ -109,3 +109,19 @@ export const exportRunsModal = (runsOverviewModel, modalHandler) => {
         }),
     ]);
 };
+
+/**
+ * Builds a button which will open popover for data export
+ * @param {RunsOverviewModel} runsModel runs overview model
+ * @param {ModelModel} modalModel modal model
+ * @returns {Component} button
+ */
+export const exportRunsTriggerAndModal = (runsModel, modalModel) => h('button.btn.btn-primary.w-15.h2.mlauto#export-runs-trigger', {
+    disabled: runsModel.runs.match({
+        Success: (payload) => payload.length === 0,
+        NotAsked: () => true,
+        Failure: () => true,
+        Loading: () => true,
+    }),
+    onclick: () => modalModel.display({ content: (modalModel) => exportRunsModal(runsModel, modalModel), size: 'medium' }),
+}, 'Export Runs');
