@@ -14,6 +14,7 @@
 import { Observable } from '/js/src/index.js';
 import { RunDetailsModel } from './Details/RunDetailsModel.js';
 import { RunsOverviewModel } from './Overview/RunsOverviewModel.js';
+import { RunsPerLhcPeriodOverviewModel } from './RunPerPeriod/RunsPerLhcPeriodOverviewModel.js';
 
 /**
  * Model representing handlers for runs page
@@ -31,6 +32,8 @@ export class RunsModel extends Observable {
         this._detailsModel.bubbleTo(this);
         this._overviewModel = new RunsOverviewModel(model);
         this._overviewModel.bubbleTo(this);
+        this._perLhcPeriodOverviewModel = new RunsPerLhcPeriodOverviewModel(model);
+        this._perLhcPeriodOverviewModel.bubbleTo(this);
     }
 
     /**
@@ -45,7 +48,7 @@ export class RunsModel extends Observable {
 
     /**
      * Returns the model for the overview page
-     * @return {LhcPeriodsOverviewModel} the overview model
+     * @return {RunsOverviewModel} the overview model
      */
     get overviewModel() {
         return this._overviewModel;
@@ -70,5 +73,24 @@ export class RunsModel extends Observable {
      */
     get detailsModel() {
         return this._detailsModel;
+    }
+
+    /**
+     * Load runs overview data
+     * @return {void}
+     */
+    loadPerLhcPeriodOverview({ lhcPeriodName }) {
+        if (!this._perLhcPeriodOverviewModel.pagination.isInfiniteScrollEnabled) {
+            this._perLhcPeriodOverviewModel.fetchDetectors();
+            this._perLhcPeriodOverviewModel.fetchRuns({ lhcPeriodName });
+        }
+    }
+
+    /**
+     * Returns the model for the overview page
+     * @return {RunsPerLhcPeriodOverviewModel} the overview model
+     */
+    get perLhcPeriodOverviewModel() {
+        return this._perLhcPeriodOverviewModel;
     }
 }
