@@ -19,12 +19,12 @@ import { paginationComponent } from '../../../components/Pagination/paginationCo
 import { filtersPanelPopover } from '../../../components/Filters/common/filtersPanelPopover.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
 
-const TABLEROW_HEIGHT = 35;
+const TABLEROW_HEIGHT = 42;
 // Estimate of the navbar and pagination elements height total; Needs to be updated in case of changes;
 const PAGE_USED_HEIGHT = 215;
 
 /**
- * Render LHC Periods overview model
+ * Render LHC Periods overview page
  * @param {Model} model The overall model object.
  * @returns {Component} The overview screen
  */
