@@ -16,13 +16,9 @@ const { OpenIdConfig, JwtConfig } = require('../config');
 const { HttpServer } = require('@aliceo2/web-ui');
 const path = require('path');
 const buildEndpoints = require('./routers');
-const {
-    user: {
-        CreateOrUpdateUser,
-    },
-} = require('../usecases');
 const { BkpRoles } = require('../domain/enums/BkpRoles.js');
 const { isInTestMode } = require('../utilities/env-utils');
+const { createOrUpdateUser } = require('./services/user/createOrUpdateUser.js');
 
 /**
  * WebUI implementation of the Server.
@@ -61,7 +57,14 @@ class WebUiServer {
                     return;
                 }
 
-                const user = await new CreateOrUpdateUser().execute(request.session);
+                const user = await createOrUpdateUser(request.session);
+                if (request.session.access) {
+                    user.access = request.session.access;
+                }
+                if (request.session.username) {
+                    user.username = request.session.username;
+                }
+
                 if (isInTestMode() && request.query?.token) {
                     user.access = user.access ?? request.query.token === 'admin' ? [BkpRoles.ADMIN] : [BkpRoles.GUEST];
                 }
